+++ sinks/signalfx/signalfx.go
@@ -17,7 +17,7 @@ import (
 	"github.com/stripe/veneur/trace"
 )
 
-type SignalFXSink struct {
+type SignalFxSink struct {
 	client           dpsink.Sink
 	endpoint         string
 	hostnameTag      string
@@ -28,8 +28,8 @@ type SignalFXSink struct {
 	traceClient      *trace.Client
 }
 
-// NewSignalFXSink creates a new SignalFX sink for metrics.
-func NewSignalFXSink(apiKey string, endpoint string, hostnameTag string, hostname string, commonDimensions map[string]string, stats *statsd.Client, log *logrus.Logger, client dpsink.Sink) (*SignalFXSink, error) {
+// NewSignalFxSink creates a new SignalFx sink for metrics.
+func NewSignalFxSink(apiKey string, endpoint string, hostnameTag string, hostname string, commonDimensions map[string]string, stats *statsd.Client, log *logrus.Logger, client dpsink.Sink) (*SignalFxSink, error) {
 	if client == nil {
 		httpSink := sfxclient.NewHTTPSink()
 		httpSink.AuthToken = apiKey
@@ -38,7 +38,7 @@ func NewSignalFXSink(apiKey string, endpoint string, hostnameTag string, hostnam
 		client = httpSink
 	}
 
-	return &SignalFXSink{
+	return &SignalFxSink{
 		client:           client,
 		endpoint:         endpoint,
 		hostnameTag:      hostnameTag,
@@ -50,18 +50,18 @@ func NewSignalFXSink(apiKey string, endpoint string, hostnameTag string, hostnam
 }
 
 // Name returns the name of this sink.
-func (sfx *SignalFXSink) Name() string {
+func (sfx *SignalFxSink) Name() string {
 	return "signalfx"
 }
 
 // Start begins the sink. For SignalFx this is a noop.
-func (sfx *SignalFXSink) Start(traceClient *trace.Client) error {
+func (sfx *SignalFxSink) Start(traceClient *trace.Client) error {
 	sfx.traceClient = traceClient
 	return nil
 }
 
 // Flush sends metrics to SignalFx
-func (sfx *SignalFXSink) Flush(ctx context.Context, interMetrics []samplers.InterMetric) error {
+func (sfx *SignalFxSink) Flush(ctx context.Context, interMetrics []samplers.InterMetric) error {
 	span, _ := trace.StartSpanFromContext(ctx, "")
 	defer span.ClientFinish(sfx.traceClient)
 
@@ -98,13 +98,13 @@ func (sfx *SignalFXSink) Flush(ctx context.Context, interMetrics []samplers.Inte
 	}
 	sfx.statsd.TimeInMilliseconds("flush.total_duration_ns", float64(time.Since(flushStart).Nanoseconds()), []string{"plugin:signalfx"}, 1.0)
 	// TODO Fix these metrics to be per-metric sink
-	sfx.log.WithField("metrics", len(interMetrics)).Info("Completed flush to SignalFX")
+	sfx.log.WithField("metrics", len(interMetrics)).Info("Completed flush to SignalFx")
 
 	return err
 }
 
 // FlushEventsChecks sends events to SignalFx. It does not support checks. It is also currently disabled.
-func (sfx *SignalFXSink) FlushEventsChecks(ctx context.Context, events []samplers.UDPEvent, checks []samplers.UDPServiceCheck) {
+func (sfx *SignalFxSink) FlushEventsChecks(ctx context.Context, events []samplers.UDPEvent, checks []samplers.UDPServiceCheck) {
 	span, _ := trace.StartSpanFromContext(ctx, "")
 	defer span.ClientFinish(sfx.traceClient)